third_party/WebKit/LayoutTests/platform/win/http/tests/push_messaging/subscribe-failure-worker-not-activated-expected.txt

This is a testharness.js-based test.
PASS subscribe() is rejected when the service worker is not active yet 
Harness: the test ran to completion.
